Adds base href and fixes react router index links
