+++ index.html
@@ -13,6 +13,7 @@
     <meta property="og:description" content="A powerful set of tools for managing inline styles on React elements." />
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="icon" href="./static/favicon.ico" type="image/x-icon">
+    <base href="/" />
     <!-- Fonts  -->
     <link href="https://fonts.googleapis.com/css?family=Source+Code+Pro:400,600" rel="stylesheet" type="text/css">
     <link href="/static/fonts.css" rel="stylesheet" type="text/css">
