+++ src/components/entry.jsx
@@ -1,24 +1,42 @@
 import React from "react";
 import { render } from "react-dom";
-import { Router, useRouterHistory, browserHistory, applyRouterMiddleware } from "react-router";
+import { renderToString } from "react-dom/server";
+import { Router, RouterContext, useRouterHistory, match, applyRouterMiddleware } from "react-router";
+import { createHistory, createMemoryHistory } from "history";
 import useScroll from "react-router-scroll";
+
+import Index from "../../templates/index.hbs";
 import routes from "../routes";
-import { useBasename } from "history";
-import createBrowserHistory from "history/lib/createBrowserHistory";
 
-let history = browserHistory;
+const routing = {
+  base: process.env.NODE_ENV === "production" ? "/open-source/radium" : ""
+};
 
-if (process.env.NODE_ENV === "production") {
-  history = useRouterHistory(useBasename(createBrowserHistory))({
-    basename: "/open-source/radium"
-  });
+// Client render (optional):
+// `static-site-generator-webpack-plugin` supports shimming browser globals
+// so instead of checking whether the document is undefined (always false),
+// Check whether it’s being shimmed
+if (typeof window !== "undefined" && window.__STATIC_GENERATOR !== true) { //eslint-disable-line no-undef
+  const history = useRouterHistory(createHistory)({ basename: routing.base });
+  render(
+    <Router
+      history={history}
+      routes={routes}
+      render={applyRouterMiddleware(useScroll())}
+    />,
+    document.getElementById("content")
+  );
 }
 
-render(
-  <Router
-    history={history}
-    routes={routes}
-    render={applyRouterMiddleware(useScroll())}
-  />,
-  document.getElementById("content")
-);
+// Exported static site renderer:
+export default (locals, callback) => {
+  const history = createMemoryHistory();
+  const location = history.createLocation(locals.path);
+  match({ routes, location }, (error, redirectLocation, renderProps) => {
+    callback(null, Index({
+      content: renderToString(<RouterContext {...renderProps} />),
+      bundleJs: locals.assets.main,
+      baseHref: `${routing.base}/`
+    }));
+  });
+};
